03.03.24/b:sprint3/step-1/v:0.1.8.5 feature:logout logic, refactor: replaced auth/login/logout logic back to utils API;
