+++ src/declarations.d.ts
@@ -4,3 +4,5 @@ declare module "*.svg" {
     const content: string;
     export default content;
 }
+
+declare module "*.png";
